pyproject.toml
[tool.poetry]
name = "marizaapp"
version = "0.1.0"
description = ""
authors = ["Arthur Briquet <mdpbriar@henallux.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
solara = "^1.31.0"
pandas = "^2.2.2"
plotly = "^5.21.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
